pyproject.toml
[project]
name = "clawdius"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic",
    "discord.py",
    "pydantic-settings",
    "loguru",
    "python-dotenv",
    "aiofiles",
    "playwright",
    "beautifulsoup4",
    "html2text",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
